Fix import for partial exports
The previous implementation assumed that each JSON object would have an
"SMS" key and a "MMS" key, and in that order. This caused an exception
when it tried to read the next NAME token because the next token is
actually the closing bracket of the object. This commit fixes this issue
by checking that the next token is actually a NAME. If it is, we consume
the token and handle it according to its value, which may be either
"sms" or "mms". If it's neither of those, we skip it.

Fixes #646